+++ tests/test_executor.py
@@ -1,8 +1,14 @@
 """Clean executor's tests."""
+from io import StringIO
+
 import pytest
 import redis
+from mock import mock
 
-from pytest_redis.executor import RedisExecutor
+from pytest_redis.executor import (
+    RedisExecutor, RedisMisconfigured, compare_version,
+    extract_version, RedisUnsupported
+)
 from pytest_redis.factories import get_config
 from pytest_redis.port import get_port
 
@@ -82,3 +88,72 @@ def test_redis_exec(request, parameter, config_option, value):
 def test_convert_bool(value, redis_value):
     """Check correctness of the redis_bool method."""
     assert RedisExecutor._redis_bool(value) == redis_value
+
+
+@pytest.mark.parametrize('version', (
+    u'Redis server version 2.4.14 (e9935407:0)',
+    u'Redis server version 2.4.13 (e0935407:0)'
+    u'Redis server version 2.5.0 (e9035407:0)'
+    u'Redis server version 2.3.10 (e9933407:0)'
+))
+def test_old_redis_version(request, version):
+    """Test how fixture behaves in case of old redis version."""
+    config = get_config(request)
+    with mock.patch(
+            'os.popen',
+            lambda *args: StringIO(version)
+    ):
+        with pytest.raises(RedisUnsupported):
+            RedisExecutor(
+                config['exec'],
+                databases=4,
+                redis_timeout=config['timeout'],
+                loglevel=config['loglevel'],
+                logsdir=config['logsdir'],
+                port=get_port(None),
+                host=config['host'],
+                timeout=30,
+            ).start()
+
+
+def test_not_existing_redis(request):
+    """Check handling of misconfigured redis executable path."""
+    config = get_config(request)
+    with pytest.raises(RedisMisconfigured):
+        RedisExecutor(
+            '/not/redis/here/redis-server',
+            databases=4,
+            redis_timeout=config['timeout'],
+            loglevel=config['loglevel'],
+            logsdir=config['logsdir'],
+            port=get_port(None),
+            host=config['host'],
+            timeout=30,
+        ).start()
+
+
+@pytest.mark.parametrize("versions,result", [
+    (["2.8.18", "2.6"], 1),
+    (["2.4.14", "2.6"], -1),
+    (["2.6.0", "2.6"], 0),
+    (["3.0.0", "2.6.17"], 1),
+    (["2.6.1", "2.6.17"], -1),
+])
+def test_compare_version(versions, result):
+    """Check if comparing version returns proper comparison result."""
+    assert compare_version(versions[0], versions[1]) == result
+
+
+@pytest.mark.parametrize("text,result", [
+    ("Redis server version 2.4.14 (00000000:0)", "2.4.14"),
+    ("Redis server v=2.6.13 sha=00000000:0 malloc=jemalloc-3.3.1 bits=64",
+     "2.6.13"),
+    ("1.2.5", "1.2.5"),
+    ("Test2.0.5", "2.0.5"),
+    ("2.0.5Test", "2.0.5"),
+    ("Test", None),
+    ("m.n.a 2.4.14", "2.4.14")
+])
+def test_extract_version(text, result):
+    """Check if the version extracction works correctly."""
+    assert extract_version(text) == result