+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGELOG
 unreleased
 -------
 
+- [enhancement] Version check got moved to executor, to be run just before starting Redis Server
 - [feature] ability to configure decode_responses for redis client in command line, pytest.ini or factory argument.
 - [bugfix] set decode_responses to False, same as StrictRedis default
 - [enhancement] ability to change decode_responses value